baseline_code/requirements.txt
openvino==2025.0.0
optimum==1.24.0
optimum-intel==1.22.0
nncf==2.15.0
torch==2.6.0
transformers==4.48.3
optimum-neuron
langchain_huggingface
langchain_community
sentence-transformers
pypdf
faiss-cpu